Fix map name grouping, add code for table inversion
