+++ VSharp.ML.AIAgent/ml/mutation/utils.py
@@ -1,4 +1,11 @@
-from .classes import MutableResultMapping
+from collections import defaultdict
+
+from .classes import (
+    GameMapsModelResults,
+    MapResultMapping,
+    ModelResultsOnGameMaps,
+    MutableResultMapping,
+)
 
 
 def sort_by_reward_asc_steps_desc(
@@ -11,3 +18,19 @@ def sort_by_reward_asc_steps_desc(
         result.move_reward.ForVisitedInstructions,
         -result.steps_count,
     )
+
+
+def invert_mapping(
+    model_results_on_map: GameMapsModelResults,
+) -> ModelResultsOnGameMaps:
+    inverse_mapping: ModelResultsOnGameMaps = defaultdict(list)
+
+    for map, list_of_mutable_result_mappings in model_results_on_map.items():
+        for mutable_result_mapping in list_of_mutable_result_mappings:
+            mutable, result = (
+                mutable_result_mapping.mutable,
+                mutable_result_mapping.mutable_result,
+            )
+            inverse_mapping[mutable].append(MapResultMapping(map, result))
+
+    return inverse_mapping
